pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "intelligent-doc"
version = "0.8.0"
description = "Intelligent Document Processing Platform for research papers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Divya"},
]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Framework :: FastAPI",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "celery[redis]>=5.4.0",
    "PyMuPDF>=1.25.0",
    "pytesseract>=0.3.13",
    "pdfplumber>=0.11.0",
    "Pillow>=11.0.0",
    "google-generativeai>=0.8.0",
    "python-multipart>=0.0.17",
    "python-dotenv>=1.0.1",
    "chromadb>=0.5.0",
    "rank-bm25>=0.2.2",
    "prometheus-client>=0.21.0",
    "slowapi>=0.1.9",
    "websockets>=12.0",
    "rouge-score>=0.1.2",
]

[project.urls]
Repository = "https://github.com/divya/intelligent-doc"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]
addopts = "--strict-markers"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disable_error_code = ["type-arg", "no-any-return", "no-untyped-call", "attr-defined"]

[[tool.mypy.overrides]]
module = [
    "google.generativeai.*",
    "celery.*",
    "redis.*",
    "chromadb.*",
    "pdfplumber.*",
    "fitz.*",
    "pytesseract.*",
    "rank_bm25.*",
    "slowapi.*",
    "rouge_score.*",
    "prometheus_client.*",
]
ignore_missing_imports = true
ignore_errors = true

[[tool.mypy.overrides]]
module = "src.models.database"
ignore_errors = true
